ai-study-assistant-main-1205/backend/requirements.txt
gunicorn>=20.0.4

Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
openai>=1.0.0
baidu-aip==2.2.13
dashscope==1.25.2
pydub==0.25.1
# 数据处理
numpy==1.26.2
pandas==2.1.4

